WIP: Implementation of a new (and clean) method to generate def-use pairs for dataflow criteria
Instead of doing 3 steps :
- Count the number of def/use for each variable (including loop)
- Create the def-use pairs depending on the number of occurences before/after
- Use clean_dataflow (Dataflows module) to remove trivially infeasible sequences

We now use the same method than clean_dataflow, but to create sequences. It now
only takes 1 step which is equivalent of the 3 previous ones.
This method use the same options (-lannot-clean-exp, -lannot-fold-index...), and you can decide
to keep trivially infeasible sequences with the option -lannot-no-clean.
It's is much faster and the result is almost equivalent (some differences due to
the order of steps, and on pointers (WIP))

How does it work :
Using Dataflows module on each function, we propagate a state between statement which keeps the
lasts seen definition for each variable :
- if we see a def, remove older defs in the current state (except if
the option -lannot-no-clean is used), then add this new def to the state and return it
- if we see a use, then get all the def of this variable in the current state,
and for each one, create a new sequence which will be added later in the AST.
- Ofc, we don't want to have duplicate def-use, so we keep a record of each def
previously created to avoid creating them more than once. For example, in loops,
we'll iter at least (and max?) 2 times to get a fixpoint, so we don't want to
duplicate everything...
- Once we have visited the whole function, visit it one more time and add all
sequences in the ast.

This method is much faster, but it was not fully tested yet...